Revert "Hackathon : export script conversation segments to s3"
