+++ .env.sample
@@ -5,7 +5,3 @@ DEFAULT_EMAIL=rageshake@beta.gouv.fr
 DRY_RUN=true
 SCHEDULE_CRISP_INVALID_RAGESHAKE=60
 SCHEDULE_JOB_SLEEPY_CONVERSATION_IN_HOURS=24
-
-S3_ACCESS_KEY_ID=
-S3_BUCKET_NAME=
-S3_SECRET_ACCESS_KEY=
