+++ README.md
@@ -17,14 +17,6 @@ The following environment variables must be set before running the script:
 - `SCHEDULE_CRISP_INVALID_RAGESHAKE`: cron schedule for the job crips invalid rageshake (by default 60 seconds)
 - `SCHEDULE_JOB_SLEEPY_CONVERSATION_IN_HOURS`: cron schedule for the job job_process_sleepy_conversations (by default 24 hours)
 
-For the statistics export to the S3 OVH bucket
-- `S3_ACCESS_KEY_ID` : access key to the export bucket
-- `S3_SECRET_ACCESS_KEY` : secret access key to the export bucket
-- `S3_BUCKET_NAME` : bucket name of the export bucket
-
-Crisp plugin is defined here : https://marketplace.crisp.chat/plugins/plugin/50c492e5-7175-45d0-a449-913aa8bc6cfd/tokens/
-
-
 ## Usage
 
 To run the script, simply run the following command:
