+++ requirements.txt
@@ -1,4 +1,3 @@
 requests==2.28.2
 python-dotenv
 schedule
-boto3 = "*"
