+++ test_sleepy_conversations.py
@@ -10,7 +10,7 @@ conversation_id = "session_691ea0d5-0543-427b-85dd-95bc412ceb27"
 
 class TestFunctions(unittest.TestCase):
     
-    @unittest.skip("integration tests are skip by default, warning this send messages to real conversations")
+    @unittest.skip("warning this sends real email to real people")
     def test_job_process_sleepy_conversation(self):
         job_process_sleepy_conversations(10)
     
@@ -21,7 +21,7 @@ class TestFunctions(unittest.TestCase):
         print(f'sleepy conversations : {len(sleepy_conversations)}')
         self.assertEqual(len(sleepy_conversations), max_conversations)
 
-    @unittest.skip("integration tests are skip by default")
+    @unittest.skip("")
     def test_is_last_message_from_operator(self):
         #this conversation ends with a note from us
         conversation_id = ""
@@ -30,12 +30,12 @@ class TestFunctions(unittest.TestCase):
         self.assertEqual(is_last_message_from_operator(conversation_id), False)
 
 
-    @unittest.skip("integration tests are skip by default")
+    @unittest.skip("")
     def test_must_send_message(self):
        simple_message = "coucou"
        send_message(conversation_id, simple_message)
 
-    @unittest.skip("integration tests are skip by default")
+    @unittest.skip("")
     def test_must_send_multiline_message_with_markdown(self):
        message = """*coucou*
                     
@@ -45,7 +45,7 @@ class TestFunctions(unittest.TestCase):
                     """
        send_message(conversation_id, message)
 
-    @unittest.skip("integration tests are skip by default")
+    @unittest.skip("")
     def test_must_change_state(self):
         new_state = "pending"
         change_conversation_state(conversation_id, new_state)
@@ -53,9 +53,9 @@ class TestFunctions(unittest.TestCase):
         print(state)
         self.assertEqual(state, new_state)
 
-    @unittest.skip("integration tests are skip by default")
+    @unittest.skip("")
     def test_wakeup_sleepy_conversation(self):
-        wakeup_sleepy_conversation("session_2e6e0ca1-58cf-4937-b3e5-201cf44a652c")
+        wakeup_sleepy_conversation("")
